+++ src/main/java/pages/blazedemo/MainPage.java
@@ -1,8 +1,9 @@
-package pages;
+package pages.blazedemo;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.Select;
+import pages.BasePage;
 
 public class MainPage extends BasePage {
